+++ netaddr/tests/ip/test_ip_v4.py
@@ -106,10 +106,14 @@ def test_ipnetwork_index_operations_v4():
     ['start', 'stop', 'step'],
     [
         [None, None, -1],
+        [-1, 0, -2],
         [-1, None, -1],
+        [-1, None, -3],
         [0, None, 4],
         [0, -1, None],
         [0, 4, None],
+        [1, None, 4],
+        [1, 4, 2],
     ],
 )
 def test_ipnetwork_slice_operations_v4(start, stop, step):